pyproject.toml
[tool.poetry]
name = "webportal-portfolio"
version = "0.1.0"
description = ""
authors = ["Ravi <Ravinder@trybz.com.au>"]
license = "MIT"
readme = "README.md"
packages = [{include = "webportal"}]

[tool.poetry.dependencies]
boto3 = "^1.26.155"
channels = {extras = ["daphne"], version = "^4.0.0"}
django = "^4.2.2"
django-cors-headers = "^4.1.0"
django-filter = "^23.2"
django-split-settings = "^1.2.0"
django-storages = "^1.13.2"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
pillow = "^9.5.0"
psycopg2 = "^2.9.6"
pydantic = "^1.10.9"
pynacl = "^1.5.0"
python = "^3.10"
pyyaml = "^6.0"
requests = "^2.31.0"
whitenoise = "^6.5.0"
openai = "^1.7.2"
celery = {extras = ["redis"], version = "^5.3.6"}
channels-redis = "^4.2.0"

[tool.poetry.group.dev.dependencies]
colorlog = "^6.7.0"
pre-commit = "^3.3.3"
pytest = "^7.4.0"
pytest-xdist = "^3.3.1"
pytest-django = "^4.5.2"
model-bakery = "^1.12.0"

[tool.isort]
line_length = 119
multi_line_output = 5

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "webportal.project.settings"
django_find_project = false
python_files = "test_*.py"
testpaths = ["webportal"]
filterwarnings = "ignore::DeprecationWarning:^(?!node\\.).*:"

[tool.yapf]
align_closing_bracket_with_visual_indent = true
based_on_style = "google"
coalesce_brackets = true
column_limit = 119
dedent_closing_brackets = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
